+++ sql/mst/kruskal.sql
@@ -30,7 +30,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 CREATE OR REPLACE FUNCTION pgr_kruskal(
     TEXT,
 
-    OUT seq INTEGER,
+    OUT seq BIGINT,
     OUT edge BIGINT,
     OUT cost FLOAT)
 RETURNS SETOF RECORD AS
